+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v3.11.0-dev
 
 - Enhancements
+    - Added support to use "form factors" client hint header for improved detection
     - Default upstream database version is now `6.4.0`
     - Improved mobile app detection
     - Upstream improvements for ARM platform detection have been integrated